Added badge option to nav_item helper
If :badge option is passed, the content of it will be added to the link as a badge (useful for counts). This is a shorthand to avoid passing a block just to add the badge.

The above forced to do a refactor of nav_item, thus spanning nav_link_tag private helper method.